+++ CHANGELOG.md
@@ -4,9 +4,12 @@
 
 ### Added
 
+- Transform.Inverted()
+
 ### Removed
 
 ### Changed
+
 - Make AnalysisMesh handle single valued analysis.
 
 ### Fixed
@@ -19,6 +22,7 @@
 - `GraphicsBuffers`
 
 ### Removed
+
 - `BuiltInMaterials.Dirt`
 - `BuiltInMaterials.Grass`
 
@@ -36,7 +40,6 @@
 
 - Fixed a bug in `ConvexHull.FromPoints` when multiple X coordinates are equal.
 - Fixed a bug in `Grid2d(Polygon, Vector3, Vector3, Vector3)` where U or V directions skew slightly when they nearly parallel with a boundary edge.
-
 
 ## 0.8.5